Added Perk List
Wrote a small script to dump the contents of the Perks List from a tab
separated values spreadsheet to arrays. Added the placeholder perks list
and a method to list them all and their descriptions. Some of them might
need to be rewritten as they refer to dice-roll rules. And currently
there's no code to actually make sense of ranks and requirements, but
they import correctly.