Added reviver method to unit.
Working on serializing objects into JSON and reviving them. This is not
a complete implementation, but it at least creates a unit object from
the string. Need to make it set the Skills to created Skill objects
next, and then match everything else up with the appropriate property
(weight to weight, food to food, etc.)