Save and Load logic methods work correctly.
The logic behind the Save and Load methods now do what they're supposed
to do during the session, but do not work as between-session storage.
The amount of data used to represent one settlement with one unit is
approximately 16 - 64KB,which is about 4 times the total amount of
storage available in a cookie at the minimum. So... Cookies aren't
looking like the best solution.  Considering trying out jstorage
instead.