More Mad_Modd
Mad Modd skills are now set correctly when a new unit is created while
that is toggled on. However, it ToString will break, because it defaults
to using the normal skilllist instead of skilllistmadmodd AND the method
that applies skillpoints will also break, since it also assumes the
normal skill set. Finally, none of the Mad-Modd specific skills save or
load. Consider putting a value in a unit that tracks if it's a
MadModd/Revised rules char, and that is automatically set at creation
time based on what the Controller.useMadModdRules says. That way, when
saving or reloading, we know which skills to draw from and rebuild and
whatnot WITHOUT asking the controller again.